+++ Gribouillage.Mobile/Gribouillage/Gribouillage.csproj
@@ -19,7 +19,6 @@
 
   <ItemGroup>
     <Folder Include="Models\" />
-    <Folder Include="Common\" />
   </ItemGroup>
   <ItemGroup>
     <None Remove=".env.sample" />
